pyproject.toml
[project]
name = "crypto-etl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apache-airflow>=3.0.3",
    "pandas>=2.3.1",
    "pyarrow>=21.0.0",
    "requests>=2.32.4",
    "sqlalchemy>=1.4.54",
]
